requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.4
PyYAML==5.1.1
requests==2.22.0
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
